+++ src/node/resolver.ts
@@ -7,7 +7,7 @@ import {
   moduleRE,
   fileToRequestMap
 } from './server/serverPluginModuleResolve'
-import { OPTIMIZE_CACHE_DIR } from './depOptimizer'
+import { resolveOptimizedCacheDir } from './depOptimizer'
 import chalk from 'chalk'
 
 export interface Resolver {
@@ -65,9 +65,9 @@ export const resolveExt = (id: string) => {
     }
     const queryMatch = id.match(/\?.*$/)
     const query = queryMatch ? queryMatch[0] : ''
-    const reoslved = cleanId + inferredExt + query
-    debug(`(extension) ${id} -> ${reoslved}`)
-    return reoslved
+    const resolved = cleanId + inferredExt + query
+    debug(`(extension) ${id} -> ${resolved}`)
+    return resolved
   }
   return id
 }
@@ -115,6 +115,7 @@ export function createResolver(
   }
 }
 
+export const jsSrcRE = /\.(?:(?:j|t)sx?|vue)$|\.mjs$/
 const deepImportRE = /^([^@][^/]*)\/|^(@[^/]+\/[^/]+)\//
 
 export function resolveBareModule(root: string, id: string, importer: string) {
@@ -126,18 +127,23 @@ export function resolveBareModule(root: string, id: string, importer: string) {
   if (pkgInfo) {
     return pkgInfo[0]
   }
-  const deepMatch = id.match(deepImportRE)
-  if (deepMatch) {
-    const depId = deepMatch[1] || deepMatch[2]
-    if (resolveOptimizedModule(root, depId)) {
-      console.error(
-        chalk.yellow(
-          `\n[vite] Avoid deep import "${id}" since "${depId}" is a ` +
-            `pre-optimized dependency.\n` +
-            `Prefer importing from the module directly.\n` +
-            `Importer: ${importer}\n`
+
+  // check and warn deep imports on optimized modules
+  const ext = path.extname(id)
+  if (!ext || jsSrcRE.test(ext)) {
+    const deepMatch = id.match(deepImportRE)
+    if (deepMatch) {
+      const depId = deepMatch[1] || deepMatch[2]
+      if (resolveOptimizedModule(root, depId)) {
+        console.error(
+          chalk.yellow(
+            `\n[vite] Avoid deep import "${id}" since "${depId}" is a ` +
+              `pre-optimized dependency.\n` +
+              `Prefer importing from the module directly.\n` +
+              `Importer: ${importer}\n`
+          )
         )
-      )
+      }
     }
   }
   return id
@@ -154,8 +160,9 @@ export function resolveOptimizedModule(
     return cached
   }
 
-  if (!id.endsWith('.js')) id += '.js'
-  const file = path.join(root, OPTIMIZE_CACHE_DIR, id)
+  const cacheDir = resolveOptimizedCacheDir(root)
+  if (!cacheDir) return
+  const file = path.join(cacheDir, id)
   if (fs.existsSync(file)) {
     viteOptimizedMap.set(id, file)
     return file