+++ src/node/server/index.ts
@@ -1,4 +1,5 @@
-import http, { Server } from 'http'
+import { RequestListener, Server } from 'http'
+import { ServerOptions } from 'https'
 import Koa from 'koa'
 import chokidar from 'chokidar'
 import { createResolver, InternalResolver } from '../resolver'
@@ -14,8 +15,11 @@ import { esbuildPlugin } from './serverPluginEsbuild'
 import { ServerConfig } from '../config'
 import { createServerTransformPlugin } from '../transform'
 import { serviceWorkerPlugin } from './serverPluginServiceWorker'
+import { htmlPlugin } from './serverPluginHtml'
 import { proxyPlugin } from './serverPluginProxy'
-
+import { createCertificate } from '../utils/createCertificate'
+import fs from 'fs-extra'
+import path from 'path'
 export { rewriteImports } from './serverPluginModuleRewrite'
 
 export type ServerPlugin = (ctx: ServerPluginContext) => void
@@ -40,7 +44,7 @@ export function createServer(config: ServerConfig = {}): Server {
   } = config
 
   const app = new Koa()
-  const server = http.createServer(app.callback())
+  const server = resolveServer(config, app.callback())
   const watcher = chokidar.watch(root, {
     ignored: [/node_modules/]
   }) as HMRWatcher
@@ -67,6 +71,7 @@ export function createServer(config: ServerConfig = {}): Server {
     ...(transforms.length ? [createServerTransformPlugin(transforms)] : []),
     esbuildPlugin,
     jsonPlugin,
+    htmlPlugin,
     assetPathPlugin,
     serveStaticPlugin
   ]
@@ -81,4 +86,43 @@ export function createServer(config: ServerConfig = {}): Server {
   }) as any
 
   return server
+}
+
+function resolveServer(
+  { https = false, httpsOption = {} }: ServerConfig,
+  requestListener: RequestListener
+) {
+  if (https) {
+    return require('https').createServer(
+      resolveHttpsConfig(httpsOption),
+      requestListener
+    )
+  } else {
+    return require('http').createServer(requestListener)
+  }
+}
+
+function resolveHttpsConfig(httpsOption: ServerOptions) {
+  const { ca, cert, key, pfx } = httpsOption
+  Object.assign(httpsOption, {
+    ca: readFileIfExits(ca),
+    cert: readFileIfExits(cert),
+    key: readFileIfExits(key),
+    pfx: readFileIfExits(pfx)
+  })
+  if (!httpsOption.key || !httpsOption.cert) {
+    httpsOption.cert = httpsOption.key = createCertificate()
+  }
+  return httpsOption
+}
+
+function readFileIfExits(value?: string | Buffer | any) {
+  if (value && !Buffer.isBuffer(value)) {
+    try {
+      return fs.readFileSync(path.resolve(value as string))
+    } catch (e) {
+      return value
+    }
+  }
+  return value
 }